message for nocoupmyanmar
Hello,

We are part of the Resistance in Hong Kong, and we look at what is happening to you in Myanmar with concern and great sadness.  We share your values and stand with you in the fight for a just and free society.

We have been writing for the safety and the effectiveness of the protesters and anyone involved in the CDM movement.

So far, we have written in English about the tactics and the roles we used in HK protests. We are looking for peers who know the Myanmar language to contribute to translations.

Here is what we have so far (https://docs.google.com/document/d/1ZrIiXypVUvPIRs9JG8AsU55FkLsz81pqZstKQcbsAHc/edit#)

We would like to know if you are interested in contributing to this project to help the movement. Drop an email anonymously at (piainoise@protonmail.com) if interested in collaboration.

white-fish